move checkpoint stuff back to trainer utils
