+++ skyrl-train/skyrl_train/utils/trainer_utils.py
@@ -89,6 +89,96 @@ def extract_step_from_path(path: str) -> int:
     return -1
 
 
+def get_latest_checkpoint_step(checkpoint_base_path: str) -> int:
+    """
+    Get the latest global step from checkpoint directory by reading latest_ckpt_global_step.txt.
+
+    Args:
+        checkpoint_base_path: Base path where checkpoints are stored
+
+    Returns:
+        int: Latest global step, or 0 if no checkpoint found
+    """
+    latest_file_path = os.path.join(checkpoint_base_path, "latest_ckpt_global_step.txt")
+
+    if not io.exists(latest_file_path):
+        return 0
+
+    try:
+        content = io.read_text(latest_file_path).strip()
+        return int(content)
+    except (ValueError, IOError) as e:
+        logger.warning(f"Failed to read latest checkpoint step from {latest_file_path}: {e}")
+        return 0
+
+
+def list_checkpoint_dirs(checkpoint_base_path: str) -> list[str]:
+    """
+    List all checkpoint directories in the base path.
+
+    Args:
+        checkpoint_base_path: Base path where checkpoints are stored
+
+    Returns:
+        list[str]: List of checkpoint directory names
+    """
+    if not io.exists(checkpoint_base_path):
+        return []
+
+    try:
+        all_items = io.list_dir(checkpoint_base_path)
+        # Filter for directories that match the global_step_* pattern
+        checkpoint_dirs = []
+        for item in all_items:
+            # Get just the basename for pattern matching
+            basename = os.path.basename(item)
+            if basename.startswith("global_step_") and io.isdir(item):
+                checkpoint_dirs.append(basename)
+
+        return sorted(checkpoint_dirs)
+    except Exception as e:
+        logger.warning(f"Failed to list checkpoint directories from {checkpoint_base_path}: {e}")
+        return []
+
+
+def cleanup_old_checkpoints(checkpoint_base_path: str, max_checkpoints: int, current_step: int) -> None:
+    """
+    Clean up old checkpoints, keeping only the most recent ones.
+
+    Args:
+        checkpoint_base_path: Base path where checkpoints are stored
+        max_checkpoints: Maximum number of checkpoints to keep
+        current_step: Current global step (for logging)
+    """
+    if max_checkpoints <= 0:
+        return
+
+    checkpoint_dirs = list_checkpoint_dirs(checkpoint_base_path)
+
+    if len(checkpoint_dirs) <= max_checkpoints:
+        return
+
+    # Sort by step number (extract number from global_step_N)
+    def extract_step(dirname):
+        try:
+            return int(dirname.split("global_step_")[1])
+        except (IndexError, ValueError):
+            return 0
+
+    checkpoint_dirs.sort(key=extract_step)
+
+    # Remove oldest checkpoints
+    dirs_to_remove = checkpoint_dirs[:-max_checkpoints]
+
+    for dir_name in dirs_to_remove:
+        full_path = os.path.join(checkpoint_base_path, dir_name)
+        try:
+            io.remove(full_path)
+            step_num = extract_step(dir_name)
+            logger.info(f"Cleaned up old checkpoint: global_step_{step_num}")
+        except Exception as e:
+            logger.warning(f"Failed to remove old checkpoint {full_path}: {e}")
+
 
 def validate_consistency_for_latest_checkpoint(
     root_ckpt_folder: str, ckpt_iteration: int, checkpoint_path: str, latest_checkpoint_file: str, save_interval: int
@@ -99,7 +189,7 @@ def validate_consistency_for_latest_checkpoint(
     Otherwise, the folder state is inconsistent and the user should delete other checkpoints.
     """
     if io.exists(root_ckpt_folder):
-        checkpoint_dirs = io.list_checkpoint_dirs(root_ckpt_folder)
+        checkpoint_dirs = list_checkpoint_dirs(root_ckpt_folder)
         if checkpoint_dirs:
             global_step_values = [extract_step_from_path(d) for d in checkpoint_dirs]
             max_global_step_in_folder = max(global_step_values)
